Refactored global config import from settings.py to ddtrace/__init__.py
