+++ ddtrace/contrib/aiohttp/middlewares.py
@@ -1,5 +1,7 @@
 import asyncio
 
+from ddtrace import config
+
 from .. import trace_utils
 from ...compat import stringify
 from ...constants import ANALYTICS_SAMPLE_RATE_KEY
@@ -7,7 +9,6 @@ from ...constants import SPAN_MEASURED_KEY
 from ...ext import SpanTypes
 from ...ext import http
 from ...propagation.http import HTTPPropagator
-from ...settings import config
 from ..asyncio import context_provider
