+++ ddtrace/contrib/aiopg/connection.py
@@ -2,6 +2,7 @@ import asyncio
 
 from aiopg.utils import _ContextManager
 
+from ddtrace import config
 from ddtrace.vendor import wrapt
 
 from .. import dbapi
@@ -10,7 +11,6 @@ from ...constants import SPAN_MEASURED_KEY
 from ...ext import SpanTypes
 from ...ext import sql
 from ...pin import Pin
-from ...settings import config
 
 
 class AIOTracedCursor(wrapt.ObjectProxy):
