+++ ddtrace/contrib/algoliasearch/patch.py
@@ -1,6 +1,6 @@
+from ddtrace import config
 from ddtrace.ext import SpanTypes
 from ddtrace.pin import Pin
-from ddtrace.settings import config
 from ddtrace.utils.wrappers import unwrap as _u
 from ddtrace.vendor.wrapt import wrap_function_wrapper as _w
