+++ ddtrace/contrib/bottle/trace.py
@@ -4,13 +4,13 @@ from bottle import request
 from bottle import response
 
 import ddtrace
+from ddtrace import config
 
 from .. import trace_utils
 from ...constants import ANALYTICS_SAMPLE_RATE_KEY
 from ...constants import SPAN_MEASURED_KEY
 from ...ext import SpanTypes
 from ...propagation.http import HTTPPropagator
-from ...settings import config
 
 
 class TracePlugin(object):
