+++ ddtrace/contrib/falcon/middleware.py
@@ -1,5 +1,6 @@
 import sys
 
+from ddtrace import config
 from ddtrace.ext import SpanTypes
 from ddtrace.ext import http as httpx
 from ddtrace.propagation.http import HTTPPropagator
@@ -8,7 +9,6 @@ from .. import trace_utils
 from ...compat import iteritems
 from ...constants import ANALYTICS_SAMPLE_RATE_KEY
 from ...constants import SPAN_MEASURED_KEY
-from ...settings import config
 
 
 class TraceMiddleware(object):
