+++ ddtrace/contrib/httplib/patch.py
@@ -1,5 +1,6 @@
 import sys
 
+from ddtrace import config
 from ddtrace.vendor import six
 from ddtrace.vendor import wrapt
 
@@ -12,7 +13,6 @@ from ...ext import SpanTypes
 from ...internal.logger import get_logger
 from ...pin import Pin
 from ...propagation.http import HTTPPropagator
-from ...settings import config
 from ...utils.formats import asbool
 from ...utils.formats import get_env
 from ...utils.wrappers import unwrap as _u
