+++ ddtrace/contrib/sqlalchemy/engine.py
@@ -17,6 +17,7 @@ from sqlalchemy.event import listen
 
 # project
 import ddtrace
+from ddtrace import config
 
 from ...constants import ANALYTICS_SAMPLE_RATE_KEY
 from ...constants import SPAN_MEASURED_KEY
@@ -24,7 +25,6 @@ from ...ext import SpanTypes
 from ...ext import net as netx
 from ...ext import sql as sqlx
 from ...pin import Pin
-from ...settings import config
 
 
 def trace_engine(engine, tracer=None, service=None):
