+++ ddtrace/contrib/sqlite3/patch.py
@@ -2,13 +2,13 @@
 import sqlite3
 import sqlite3.dbapi2
 
+from ddtrace import config
 from ddtrace.vendor import wrapt
 
 # project
 from ...contrib.dbapi import FetchTracedCursor
 from ...contrib.dbapi import TracedConnection
 from ...contrib.dbapi import TracedCursor
-from ...pin import Pin
 from ...settings import config
