+++ ddtrace/contrib/tornado/handlers.py
@@ -1,11 +1,12 @@
 from tornado.web import HTTPError
 
+from ddtrace import config
+
 from ...constants import ANALYTICS_SAMPLE_RATE_KEY
 from ...constants import SPAN_MEASURED_KEY
 from ...ext import SpanTypes
 from ...ext import http
 from ...propagation.http import HTTPPropagator
-from ...settings import config
 from .constants import CONFIG_KEY
 from .constants import REQUEST_CONTEXT_KEY
 from .constants import REQUEST_SPAN_KEY
